+++ Source/OxyPlot.SkiaSharp/JpegExporter.cs
@@ -70,7 +70,7 @@ namespace OxyPlot.SkiaSharp
             using var bitmap = new SKBitmap(this.Width, this.Height);
 
             using (var canvas = new SKCanvas(bitmap))
-            using (var context = new SkiaRenderContext { RendersToScreen = false, SkCanvas = canvas })
+            using (var context = new SkiaRenderContext { RenderTarget = RenderTarget.PixelGraphic, SkCanvas = canvas })
             {
                 canvas.Clear(SKColors.White);
                 var dpiScale = this.Dpi / 96;